Refactor Prometheus metrics middleware and improve code organization:
- Move `PrometheusMiddleware` to a new `services/metrics.py` module for better modularity.
- Simplify imports and remove unused code from `api/metrics.py`.
- Update middleware imports in `app/middleware.py` to reflect the new structure.
- Minor improvements to docstrings for consistency.
