+++ backend/src/api/metrics.py
@@ -1,17 +1,6 @@
-import time
-from typing import Callable, List
-
-from fastapi import APIRouter, Request
+from fastapi import APIRouter
 from fastapi.responses import JSONResponse, Response
 from prometheus_client import CONTENT_TYPE_LATEST, generate_latest
-from starlette.middleware.base import BaseHTTPMiddleware
-
-from ..schemas.metric import (
-    EXCEPTIONS_TOTAL,
-    REQUEST_DURATION_HIST,
-    REQUESTS_IN_PROGRESS,
-    RESPONSES_TOTAL,
-)
 
 router_metrics = APIRouter(
     prefix="/api/metrics",
@@ -53,54 +42,3 @@ async def get_metrics_doc() -> Response:
     Prometheus metrics endpoint
     """
     return Response(generate_latest(), media_type=CONTENT_TYPE_LATEST)
-
-
-EXCLUDE_PATH_PREFIXES: List[str] = [
-    "/api/metrics",
-    "/api/health",
-    "/static",
-    "/docs",
-    "/openapi.json",
-]
-
-
-def is_excluded_path(path: str) -> bool:
-    return any(path.startswith(p) for p in EXCLUDE_PATH_PREFIXES)
-
-
-class PrometheusMiddleware(BaseHTTPMiddleware):
-    async def dispatch(self, request: Request, call_next: Callable) -> Response:
-        if is_excluded_path(request.url.path):
-            return await call_next(request)
-
-        route = request.scope.get("route")
-        if route and hasattr(route, "path"):
-            path_label = route.path
-        else:
-            path_label = request.url.path
-
-        method = request.method
-
-        REQUESTS_IN_PROGRESS.labels(method=method, path=path_label).inc()
-        start = time.perf_counter()
-
-        try:
-            response = await call_next(request)
-            status_code = str(response.status_code)
-
-            RESPONSES_TOTAL.labels(
-                status_code=status_code, method=method, path=path_label
-            ).inc()
-            return response
-        except Exception as exc:
-            exc_type = type(exc).__name__
-            EXCEPTIONS_TOTAL.labels(
-                exception_type=exc_type, method=method, path=path_label
-            ).inc()
-            raise
-        finally:
-            duration = time.perf_counter() - start
-            REQUEST_DURATION_HIST.labels(method=method, path=path_label).observe(
-                duration
-            )
-            REQUESTS_IN_PROGRESS.labels(method=method, path=path_label).dec()
