+++ backend/src/i18n.py
@@ -68,9 +68,6 @@ class LanguageMiddleware(BaseHTTPMiddleware):
     based on the Accept-Language header in the incoming
     request. It uses the set_locale function from the i18n
     module to determine the appropriate language for the request.
-
-    Attributes:
-        None
     """
 
     async def dispatch(self, request: Request, call_next: Callable) -> Response:
@@ -82,9 +79,9 @@ class LanguageMiddleware(BaseHTTPMiddleware):
         passes the request to the next middleware or route handler.
 
         Args:
-            request (Request): The incoming request.
+            request (Request): The incoming request
             call_next (Callable): The function to call to proceed
-            to the next middleware or route handler.
+            to the next middleware or route handler
 
         Returns:
             Response: The response generated by the next middleware
